feat: Consolidate and standardize `PixelContext`
Moves the `PixelContext` definition to the interfaces layer and replaces the existing protocol with a concrete dataclass implementation. This refactor aligns pixel generation with other pipeline variants by incorporating standard lifecycle management, status tracking, and scheduler compatibility methods.
