+++ max/python/max/interfaces/__init__.py
@@ -50,8 +50,7 @@ from .pipeline_variants import (
     EmbeddingsGenerationOutput,
     ImageContentPart,
     ImageMetadata,
-    PixelGenerationContext,
-    PixelGenerationContextType,
+    PixelContext,
     PixelGenerationInputs,
     PixelGenerationOutput,
     PixelGenerationRequest,
@@ -133,8 +132,7 @@ __all__ = [
     "PipelineTask",
     "PipelineTokenizer",
     "PipelinesFactory",
-    "PixelGenerationContext",
-    "PixelGenerationContextType",
+    "PixelContext",
     "PixelGenerationInputs",
     "PixelGenerationOutput",
     "PixelGenerationRequest",
